Track total contention time for mutexes and rwlocks
This adds the ability to track the total amount of time for which each mutex
or rwlock is contended. Effectively, this is the amount of time which is
spent blocked inside pthread_mutex_lock() and friends — we don't count time
spent inside these functions if the mutex isn't busy, however (i.e. if
pthread_mutex_trylock() doesn't return EBUSY).

It is possible for the total contention time for a given mutex to exceed the
total locked time. This can happen if the OS doesn't re-schedule a thread
which is blocked on a mutex until some time after the thread which was
holding the mutex unlocks it.

This commit replaces the “max.Time[ms]” summary column with a new
“cont.Time[ms]” column which lists the total contention time for each mutex.
I believe this is more important/useful than the maximum locked time.