Expand support for rwlocks to count read and write locks separately
All counters such as n_locked are now split into two: one for reads and one
for writes. Normal mutexes just use the write counters, whereas rwlocks now
use both the read and write counters as appropriate.

In order to display all this new data, rwlocks are now printed as two lines
in the summary table: the first is for the write counters and the second is
for the read counters. There is currently no way to print out the sum of
any two such counters, or order by this sum.