Add support for profiling condition variables
This adds a separate hash table of structures for profiling data for each
condition variable in the program. Statistics such as the number of waits,
signals, number of contended waits (waits while another thread is already
waiting), contended signals (signals while no threads are waiting) and
contention time are collected and presented in a second table in the summary
printout.